modifying the code to rely on offline file Indicies/currencies.json created offline periodically with more updated values for currency conversions.
Upon loading, the values are rounded to 5 decimal digits accuracy.

I couldn't find the syntax for jsonify the json file (adding implicitly the newlines) so left it as is for now (will be glad to learn :-) ).

This was checked with TASE scan, pls check with all/nasdaq as a safety measure.

json library is improted in sss.py